requirements.txt
sumy==0.8.1
Flask==1.1.1
numpy==1.18.1
matplotlib==3.1.3
streamlit==0.79.0
pandas==1.0.1
opencv_python_headless==4.4.0.46
gtts==2.2.2
Pillow==8.1.2
protobuf==3.15.6
PyPDF2==1.26.0
pyttsx3==2.90
